separate extra functionality into separate module
