+++ package.json
@@ -53,6 +53,8 @@
     "url": "https://github.com/phosphorjs/phosphor-widget.git"
   },
   "files": [
+    "lib/extra.js",
+    "lib/extra.d.ts",
     "lib/index.js",
     "lib/index.css",
     "lib/index.d.ts"
